Extract Bill Details/requirements.txt
Flask==2.0.1
gunicorn==20.0.4
azure-ai-formrecognizer==3.1.1